+++ doc/changelog.rst
@@ -1,4 +1,4 @@
-Shoop Boleto Change Log
+Shuup Boleto Change Log
 ===================
 
 Version 1.0.0
@@ -6,4 +6,4 @@ Version 1.0.0
 
 Released on 2016-??-??
 
-- The first version of Shoop Boleto.
+- The first version of Shuup Boleto.